Update sidebar.js
This commit enhances the `forceElementVisibility(el)` in function by adding a conditional statement `if(el){}` to prevent potential errors. With this improvement, if a new route is added that is not registered in the sidebar, it will not trigger an error. This ensures smoother functionality and prevents unnecessary interruptions in the user experience.